+++ backend/Dockerfile
@@ -23,4 +23,4 @@ COPY . .
 EXPOSE 5000
 
 # Run the command to start the Flask app
-CMD ["python", "app.py"]
+CMD ["python", "start.py"]
