+++ backend/database.py
@@ -1,0 +1,19 @@
+from sqlalchemy import create_engine, Column, Integer, String, DateTime
+from sqlalchemy.orm import declarative_base, sessionmaker
+
+# Database setup
+DATABASE_URL = "sqlite:///data.db"
+Base = declarative_base()
+
+class Event(Base):
+    __tablename__ = 'events'
+    id = Column(Integer, primary_key=True)
+    title = Column(String)
+    location = Column(String)
+    datetime = Column(DateTime)
+    channel = Column(String)
+    status = Column(String)
+
+engine = create_engine(DATABASE_URL)
+Base.metadata.create_all(engine)
+Session = sessionmaker(bind=engine)
